refactor(api): split into ferx-core / ferx-tools / ferx-cli workspace [#1114]
Make the repo a cargo workspace so model-development tooling (bootstrap,
stepwise covariate modelling, model search, cross-validation) can be written in
Rust without paying for it on every ferx-core rebuild, and without adding a
repository to the downstream pin chain.

The root stays the ferx-core package. That is load-bearing:
ferx-r/src/rust/.cargo/config.toml patches ferx-core to the repo *root*, so
moving the package under crates/ would break every local ferx-r build — and
break it silently, since the patch would simply stop applying and ferx-r would
quietly build main instead of the branch under test.

The `ferx` binary moves to crates/ferx-cli. It will grow ferx-tools subcommands,
and ferx-tools depends on ferx-core, so a binary in ferx-core would be a cycle.
The installed binary name and behaviour are unchanged; building it from source
now needs the workspace or the package named.

Enforcement (A2): api/ferx-core-public-api.txt is a committed cargo-public-api
snapshot that CI diffs, so a new `pub` item fails the build until the baseline is
updated in the same PR. The failure mode being guarded against is not ferx-tools
reaching into internals — cargo makes that impossible — but ferx-core growing an
ad-hoc `pub` in the same PR that needs it.

Deviations from the issue's sketch, each for a measured reason:

- The baseline is generated with `ci,markov,nn`, not default features. Those
  three cfgs each add public items, so a default-features baseline would let a
  `pub` behind one of them through the gate unseen.

- `-sss` (omit blanket/auto-trait/auto-derived impls). At `-s` the surface is
  9190 lines of which ~3.6k are `impl core::marker::Freeze for …`, and one new
  `pub struct` moves a dozen lines. A2 exists so a new item is one reviewable
  line; `-sss` gives exactly that (4280 items).

- tests/public_api_boundary.rs bans `#[doc(hidden)]` outright. Verified
  directly: `#[doc(hidden)] pub fn probe()` added to src/lib.rs leaves the
  baseline check GREEN, while the same fn without the attribute fails it with
  one added line. cargo-public-api omits hidden items, so A3 is not enforceable
  by the baseline alone — the attribute is a working bypass. The same file pins
  A1, since a dev-dependency on ferx-cli would not be a cycle cargo rejects.

- coverage-pr becomes `--workspace --features ferx-core/ci` rather than gaining
  a standalone member test job. Verified that cargo llvm-cov does NOT go
  workspace-wide on its own here (102 SF: records, none under crates/), so
  root-scoping would have dropped the CLI's ~960 previously-measured lines from
  the `fast` flag — lowering project coverage for a pure file move and failing a
  later crates/-only PR at 0% on its own patch gate (the #293 trap). One
  instrumented build now runs and measures the members: no third cache key, no
  second ferx-core compile.

- release.yml needs no repoint; it builds nothing, it only generates release
  notes.

Feature flags belong to ferx-core, so member-scoped and workspace-wide commands
write them package-qualified (`--features ferx-core/ci`); a bare `--features ci`
fails with "none of the selected packages contains this feature". Qualifying
them also keeps the members linking the rlib the previous CI step just built
instead of forcing a second compile under a different feature hash.

`resolver = "2"` preserves behaviour rather than changing it: a standalone
edition-2021 package gets v2 implicitly, but inside a workspace the root
package's edition no longer implies it and cargo would fall back to v1.

Verified by hand, since the failure is silent: a ferx-r build patched at this
checkout resolves ferx-core to it (`source: None`) and produces 9 diagnostics
identical to the same build patched at main — pre-existing ferx-r drift
(kappa_weights, residual_correlations, discrete_rows), untouched by the split.
Also verified: cargo package normalizes the [workspace] table away and excludes
crates/, so the crates.io path is unaffected; `cargo build --release --workspace`
produces a behaviourally identical `ferx`; the full workspace suite is green
(159 test binaries, 0 failures).
