+++ docs/development/sdlc.qmd
@@ -265,7 +265,7 @@ after each major feature) and not yet on GitHub — a known gap to formalize.
 
 | Workflow | Trigger | Jobs |
 |----------|---------|------|
-| `ci.yml` | PR + push to `main` | `Check` (compile-only, incl. the `survival` / `markov` / `slow-tests` feature combos), `Tests + coverage (core)` (`--tests --features ci`, instrumented — runs the Tier-1/2 suite *and* produces the patch-coverage report), `Tests + coverage (TTE/CTMM endpoints)` (`--features ci,markov`, only the endpoint test binaries), `Clippy`, `Format`; **`Tests + coverage (full, slow-tests)`** on weekly schedule / manual |
+| `ci.yml` | PR + push to `main` | `Check` (compile-only, incl. the `survival` / `markov` / `slow-tests` feature combos), `Tests + coverage (core)` (`--tests --features ci`, instrumented — runs the Tier-1/2 suite *and* produces the patch-coverage report), `Tests + coverage (TTE/CTMM endpoints)` (`--features ci,markov`, only the endpoint test binaries), `Clippy`, `Format`, `Public API baseline` (diffs `api/ferx-core-public-api.txt`); **`Tests + coverage (full, slow-tests)`** on weekly schedule / manual |
 | `slow-tests.yml` | nightly 06:00 UTC, manual, push to `main` touching estimation paths | Tier-3 fits to convergence (`--features ci,slow-tests --release`, `--no-fail-fast`) |
 | `docs.yml` | push to `main` touching `docs/**` | `quarto render` → deploy to `gh-pages` |
 | `release.yml` | tag `v*` or manual | GitHub release with generated notes |
@@ -326,6 +326,35 @@ cache warm. `rust-toolchain.toml` still pins nightly as the default, and `Check`
 - **Keep the build toolchain-light.** Standard-Rust users (Mac / Linux / Windows)
   and CI install ferx with `cargo build` alone. Do not reintroduce a dependency on
   a custom compiler or a non-default Cargo feature for the core gradient path.
+
+---
+
+### 8.7 The public-API gate
+
+The repo is a cargo workspace: the root package is the `ferx-core` **library**,
+with `crates/ferx-tools` (multi-fit tooling — bootstrap, covariate search,
+cross-validation) and `crates/ferx-cli` (the `ferx` binary) stacked on top. The
+boundary rule is *`ferx-core` = one model, one dataset, one fit; `ferx-tools` =
+many fits*. Statistical kernels stay in core; orchestration of repeated fits does
+not.
+
+Members can only reach `ferx-core` through its `pub` API, so the way that
+boundary erodes is core growing an ad-hoc `pub` in the same PR that needs it.
+`api/ferx-core-public-api.txt` is a committed `cargo public-api` snapshot and the
+`Public API baseline` job diffs it, which turns every widening into one explicit
+reviewable line rather than something buried in a large change. It doubles as
+semver protection for the crates.io release and for ferx-r.
+
+```bash
+tools/update-public-api.sh          # regenerate after an intended change
+tools/update-public-api.sh --check  # what CI runs
+```
+
+`#[doc(hidden)]` is banned outright (`tests/public_api_boundary.rs`): `cargo
+public-api` omits such items, so the attribute is a working bypass of this gate,
+not a style question. And anything `ferx-tools` needs must also be reachable from
+ferx-r, which consumes `ferx-core` as an ordinary external crate — an API that
+only makes sense "because we're in the same workspace" is the wrong API.
 
 ---
 
@@ -397,6 +426,10 @@ A bare `cargo update` unpins the patch and CI fails with
 
 **Checklist when changing a `pub` API or `FitResult`/sdtab field**
 
+- [ ] Regenerate the public-API baseline in the same PR
+      (`tools/update-public-api.sh`) and state which caller needs each added
+      item — the `Public API baseline` job fails until you do. See
+      [§8.7](#87-the-public-api-gate).
 - [ ] Open the matching ferx-r PR and record it in the PR cross-repo table.
 - [ ] Bump `Cargo.lock` in ferx-r via `update-ferx-core-lock.sh`.
 - [ ] Confirm the R package compiles against the new ferx-core.
@@ -425,7 +458,7 @@ A change is done only when **all** of the following hold:
 - [ ] Implements the issue's stated intent; design choices discussed where open.
 - [ ] Tests added at the correct tier; whole fast suite still < 10 min.
 - [ ] NONMEM comparison included for any numerical behaviour.
-- [ ] CI green: `check`, `test`, `survival`, `clippy`, `fmt`; coverage not degraded.
+- [ ] CI green: `check`, `test`, `survival`, `clippy`, `fmt`, `public-api`; coverage not degraded.
 - [ ] AI review run (and human review for large/numerical/DSL/design PRs).
 - [ ] Docs updated (`docs/**`, examples) for any user-visible change.
 - [ ] ferx-core ↔ ferx-r in sync (R package compiles; `Cargo.lock` bumped if needed).
